Move bbcode class in forums.php to own file
